fixed incorrect paths after moving directory
