+++ client/questions_answers/controllers.js
@@ -1,5 +1,5 @@
 import $ from 'jquery';
-import TOKEN from '../../../config.js';
+import TOKEN from '../../config.js';
 
 
 // THESE FUNCTIONS WILL BE REFACTORED TO MAKE ALL REQUESTS TO THE SERVER INSTEAD OF THE API
